Fix alexnet.csv not found error
==============================

The alexnet.csv file is in the /topologies/conv_nets path. But In the
code the the vallue of the `network` flag is: 'topologies/alexnet.csv'
.
Because of this incorrect path, when you ru scale.py file, you will get
following error:
	param_file = open(topology_file, 'r')
	FileNotFoundError: [Errno 2] No such file or directory: './topologies/alexnet.csv'
I've fixed this problem by changing this path to the correct one.